Split bridges.test.ts: move bootBridges suite to bridges.boot.test.ts (warren-5f55)
